requirements.txt
# Core FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.22
alembic==1.12.1
psycopg2-binary==2.9.10  # Windows: dùng bản prebuilt

# Redis
redis==5.0.1

# Authentication & Security
python-jose==3.3.0  # Windows: bản pip chuẩn
passlib[bcrypt]==1.7.4  # Windows: pip chuẩn
python-multipart==0.0.6

# Configuration & Environment
pydantic>=2.4,<3
pydantic-core>=2.4,<3
pydantic-settings==2.0.3
python-dotenv==1.0.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Utilities
email-validator==2.1.0

PyJWT==2.10.1
python-dateutil==2.9.0post0 
